Dashboard/requirements.txt
matplotlib==3.7.1
pandas==1.5.3
lime==0.2.0.1
seaborn==0.12.2
Pillow==9.4.0
requests==2.28.1
matplotlib==3.7.1
streamlit==1.22.0